refactor carstore meta into a class
